Update perp.$symbol.tsx
Add a script to move BGSC to the top of the list. This implementation is purely front-end: it checks each tbody, and if it contains BGSC, it moves it to the top of the table. It was implemented this way because:

I tried implementing it with dataAdapter, which works fine but only when the list is unsorted. By default, the list is sorted by volume, and in order to see the result we would have to manually click the volume sorting twice (from descending → ascending → unsorted). This does not fit Aden’s expectations.

In the SDK, the list is rendered with SideMarketsWidget and eventually handled by marketsList.widget, which takes initialSort and sorting as parameters. However, these cannot be passed directly from tradingPage. Therefore, we cannot change the default sorting mechanism at the SDK level.

Reference:
https://github.com/OrderlyNetwork/js-sdk/blob/7ac6321a2433bcfb97572da9019c7dfd8e7572b2/packages/trading/src/pages/trading/trading.ui.desktop.tsx#L97

https://github.com/OrderlyNetwork/js-sdk/blob/main/packages/markets/src/components/marketsList/marketsList.widget.tsx#L29